RIP-622, if devAuth then option to change Canvas Site ID in session
